% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_test.R
\name{MAP_test}
\alias{MAP_test}
\title{MAP test}
\usage{
MAP_test(est_result, dd = NULL, Type = c(1:5), nn = 6000)
}
\arguments{
\item{est_result}{estimation_fun result:\cr
A List contains data_use (the data information) and result1 (Parameters estimations)}

\item{dd}{True index to check the true FDR}

\item{Type}{Type of hypothesis
\itemize{
\item Type = 1 general DE detection
\item Type = 2 PDE with no time-by-treatment and NPDE with both treatment and time-by-treatment
\item Type = 3 NPDE with or without time-by-treatment
\item Type = 4 PDE with no time-by-treatment
\item Type = 5 NPDE with only time-by-treatment
}}

\item{nn}{Number of QMC nodes used to estimate the test statistics}
}
\value{
A list of results for DE analysis.\cr
\itemize{
\item Type Type of hypothesis of interest
\item FDR_final FDR estimation of each hypothesis
\item FDR_real The true FDR if dd is known
\item power Power if dd is known
\item T_x Likelihood result for each component
\item test_stat Test statistics for each hypothesis
\item ct_final_all Test statistics cut-off value at each estimated FDR level
}
}
\description{
MAP test
}
\details{
The vector of read counts for gene g, treatment group i, replicate j,
at time point \eqn{t,Y_{gij}(t)}, follows a Negative Binomial distribution
parameterized mean \eqn{\lambda_{gi}} and \eqn{\phi_g}, where
\eqn{E[Y_{gij}(t)] = lambda_{gi}(t)}.
\eqn{\lambda_{gi}(t)} is further modeled as
\eqn{\lambda_{gi}(t) = S_{ij} exp[\eta_{g1}I_{i = 2} + B'(t)\eta_{g2}I_{i = 2} + B'(t)\tau_{g}]}
We have B'(t) are design matrix, which is constructed by 2 orthorgonal polynomial bases.
\itemize{
\item t = 1,..., n_treat (or n_control if control group);
\item j = 1,..., n_rep;
\item g = 1,...,G; and
\item \eqn{[\eta_{g1}, \eta_{g2}, \tau_{g}]} ~ 4-component gausssian mixture model.
After the parameter estimation we run the MAP test to detect the DE genes.
}
}
\examples{
aaa <- proc.time()
G <- 100
k_real <- 4
p_k_real <- c(0.7, 0.1, 0.1, 0.1)
dd = rep(c(0:(k_real-1)), p_k_real * G)
result <- MAP_test(est_result = est_result, Type = c(1:5), dd = dd, nn = 3000)
aaa1 <- proc.time()
aaa1 - aaa




}
