% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPTest.R
\docType{package}
\name{MAPTest-package}
\alias{MAPTest}
\alias{MAPTest-package}
\title{MAPTest: MAP Test for detection DE genes}
\description{
MAP test for analyzing temporal gene expression data.
}
\details{
This package contains functions for statistical calculations and multiple testing procedures.
}
\author{
Meng Cao <meng.cao@colostate.edu>, \cr
F.Jay Breidt <FJay.Breidt@colostate.edu>, \cr
Wen Zhou <riczw@stat.colostate.edu>
}
\keyword{internal}
