% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_MAP.R
\name{Summary_MAP}
\alias{Summary_MAP}
\title{Summary for MAP Test}
\usage{
Summary_MAP(test_result, alpha = 0.05)
}
\arguments{
\item{test_result}{MAP_test result:\cr
A list of results for DE analysis}

\item{alpha}{Nominal level of FDR}
}
\value{
A list of results for DE analysis.\cr
\itemize{
\item Type Type of hypothesis
\item Reject_index DE genes list
\item FDR_hat Estimated FDR
}
}
\description{
Summary for MAP Test
}
\details{
The vector of read counts for gene g, treatment group i, replicate j,
at time point \eqn{t,Y_{gij}(t)}, follows a Negative Binomial distribution
parameterized mean \eqn{\lambda_{gi}} and \eqn{\phi_g}, where
\eqn{E[Y_{gij}(t)] = lambda_{gi}(t)}.
\eqn{\lambda_{gi}(t)} is further modeled as
\eqn{\lambda_{gi}(t) = S_{ij} exp[\eta_{g1}I_{i = 2} + B'(t)\eta_{g2}I_{i = 2} + B'(t)\tau_{g}]}
We have B'(t) are design matrix, which is constructed by 2 orthorgonal polynomial bases.
\itemize{
\item t = 1,..., n_treat (or n_control if control group);
\item j = 1,..., n_rep;
\item g = 1,...,G; and
\item \eqn{[\eta_{g1}, \eta_{g2}, \tau_{g}]} ~ 4-component gausssian mixture model.
After the parameter estimation we run the MAP test to detect the DE genes. At given
nominal level, a list of differential genes are returned at a given hypothesis.
}
}
\examples{
aaa <- proc.time()
G <- 100
k_real <- 4
p_k_real <- c(0.7, 0.1, 0.1, 0.1)
dd = rep(c(0:(k_real-1)), p_k_real * G)
result <- MAP_test(est_result = est_result, Type = c(1:5), dd = dd, nn = 3000)
aaa1 <- proc.time()
aaa1 - aaa
Summary_MAP(result)
}
