% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP_test.R
\name{MAP_test}
\alias{MAP_test}
\title{MAP test}
\usage{
MAP_test(est_result, dd = NULL, Type = c(1:6), nn = 6000)
}
\arguments{
\item{est_result}{estimation_fun result:\cr
A List contains data_use (the data information) and result1 (Parameters estimations)}

\item{dd}{True index to check the true FDR}

\item{Type}{Type of hypothesis
\itemize{
\item Type = 1 general DE detection
\item Type = 2 PDE with no time-by-treatment and NPDE with both treatment and time-by-treatment
\item Type = 3 NPDE with or without time-by-treatment
\item Type = 4 PDE with no time-by-treatment
\item Type = 5 NPDE with only time-by-treatment
\item Type = 6 NPDE with both treatment and time-by-treatment
}}

\item{nn}{Number of QMC nodes used to estimate the test statistics}
}
\value{
A list of results for DE analysis.\cr
\itemize{
\item Type Type of hypothesis of interest
\item FDR_final FDR estimation of each hypothesis
\item FDR_real The true FDR if dd is known
\item power Power if dd is known
\item T_x Likelihood result for each component
\item test_stat Test statistics for each hypothesis
\item ct_final_all Test statistics cut-off value at each estimated FDR level
}
}
\description{
MAP test
}
\details{
The vector of read counts for gene g, treatment group i, replicate j,
at time point \eqn{t,Y_{gij}(t)}, follows a Negative Binomial distribution
parameterized mean \eqn{\lambda_{gi}} and \eqn{\phi_g}, where
\eqn{E[Y_{gij}(t)] = \lambda_{gi}(t)}.
\eqn{\lambda_{gi}(t)} is further modeled as
\eqn{\lambda_{gi}(t) = S_{ij} \exp[\eta_{g1}I_{i = 2} + B'(t)\eta_{g2}I_{i = 2} + B'(t)\tau_{g}].}
We have \eqn{B'(t)} are design matrix, which is constructed by 2 orthorgonal polynomial bases.
\itemize{
\item t = 1,..., n_treat (or n_control if control group);
\item j = 1,..., n_rep;
\item g = 1,...,G; and
\item \eqn{[\eta_{g1}, \eta_{g2}, \tau_{g}]} ~ 4-component gausssian mixture model.
After the parameter estimation we run the MAP test to detect the DE genes.
}
}
\examples{
library(matlib)
n_basis = 2
n_control = 10
n_treat   = 10
n_rep = 3
tt_treat  = c(1:n_treat)/n_treat
nt = length(tt_treat)
ind_t = sort(sample(c(1:nt), n_control))
tt = tt_treat[ind_t]
tttt = c(rep(tt, n_rep), rep(tt_treat, n_rep))
z = x = matrix(0, length(tttt), n_basis)
z[,1] = 1.224745*tttt
z[,2] = -0.7905694 + 2.371708*tttt^2
x[,1] = z[,1] - Proj(z[,1], rep(1, length(tttt)))
x[,2] = z[,2] - Proj(z[,2], rep(1, length(tttt))) - Proj(z[,2], x[,1])
Y1 = data_generation(G = 100,
                     n_control = n_control,
                     n_treat   = n_treat,
                     n_rep     = n_rep,
                     k_real = 4,
                     sigma2_r = rep(1, 2),
                     sigma1_2_r = 1,
                     sigma2_2_r = c(3,2),
                     mu1_r = 4,
                     phi_g_r = rep(1,100),
                     p_k_real = c(0.7, 0.1, 0.1, 0.1),
                     x = x)
aaa <- proc.time()
est_result <- estimation_fun(n_control = n_control,
                             n_treat = n_treat,
                             n_rep = n_rep,
                             x = x,
                             Y1 = Y1,
                             nn = 300,
                             k = 4,
                             phi = NULL,
                             type = 2,
                             tttt = tttt)
aaa1 <- proc.time()
aaa1 - aaa

G <- 100
k_real <- 4
p_k_real <- c(0.7, 0.1, 0.1, 0.1)
dd = rep(c(0:(k_real-1)), p_k_real * G)
result = MAP_test(est_result = est_result, Type = c(1:6), dd = dd, nn = 300)
}
