% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPTest.R
\docType{package}
\name{MAPTest-package}
\alias{MAPTest}
\alias{MAPTest-package}
\title{MAPTest: MAP Test for detection DE genes}
\description{
MAP test for analyzing temporal gene expression data.
}
\details{
MAP test for analyzing temporal gene expression data.
}
\author{
\strong{Maintainer}: Meng Cao \email{meng.cao@colostate.edu}

}
\keyword{internal}
